CW2 Added CNItem processing

+++ server/src/main/java/com/remarkablesystems/childwatch/domain/food/projection/MealFoodItemFull.java
@@ -1,7 +1,5 @@
 package com.remarkablesystems.childwatch.domain.food.projection;
 
-import java.math.BigDecimal;
-
 import org.springframework.data.rest.core.config.Projection;
 
 import com.remarkablesystems.childwatch.domain.food.AgeGroup;
@@ -13,7 +11,7 @@ public interface MealFoodItemFull {
 
 	String getId();
 	AgeGroup getAgeGroup();
-	BigDecimal getQuantity();
+	double getQuantity();
 	MealFull getMeal();
 	FoodItemFull getFoodItem();
 	UnitOfMeasure getUnit();
